Prepare release of version 4.0.0-OS14 (#43)
* RMET-2119 InAppBrowser Plugin - Check for element before obtaining it (#40)

* fix: check if there's a next token before trying to obtain it

References: https://outsystemsrd.atlassian.net/browse/RMET-2119

* chore: update changelog

* RMET-2802 InAppBrowser Plugin - Multiple minor fixes from the community (#42)

* Fix memory leak - initially reported in https://github.com/apache/cordova-plugin-inappbrowser/issues/290

Whenever closing an InAppBrowser instance, a webview was left in memory with about:blank page.
This change fixes the issue by destroying and freeing the inAppWebView object.

* Fix beforeLoad not being called in some requests - initially reported in https://github.com/apache/cordova-plugin-inappbrowser/issues/686

The waitForBeforeload flag was preventing beforeLoad from being called on every GET request.

* Do not lose callbackContext when opening a SYSTEM url

This fixes a condition where it was not possible to open a SYSTEM url while an InAppBrowser instance is displayed. The callbackContext of the InAppBrowser instance was lost when the SYSTEM url was opened. This fixes the issue by not setting the callbackContext on SYSTEM urls.

* Fix crash when pausing/resuming application after closing InAppBrowser

* Fix _loadAfterBeforeload callback not working due to this.rootName not being defined.

* Reset beforeload variable if a new instance of InAppBrowser is opened without beforeload setting

* chore: update changelog

References: https://outsystemsrd.atlassian.net/browse/RMET-2802

* refactor: remove empty lines

* refactor: use ternary operator and remove unnecessary ones

Why: We can refactor the if-else block that assigns a value to beforeload to use a ternary operator. On the other hand, we can remove the unnecessary ternary operators used to assign a boolean value. They are redundant.

References: https://outsystemsrd.atlassian.net/browse/RMET-2802

---------

Co-authored-by: Leonardo Monteiro Fernandes <leonardo.monteiro.fernandes@gmail.com>
Co-authored-by: Nelson Lopes Silva <5671236+nflsilva@users.noreply.github.com>

---------

Co-authored-by: Leonardo Monteiro Fernandes <leonardo.monteiro.fernandes@gmail.com>
Co-authored-by: Nelson Lopes Silva <5671236+nflsilva@users.noreply.github.com>